Делит код проекта на модули.
