+++ js/comparison.js
@@ -1,0 +1,1 @@
+//TODO: Модуль для сравнения объявлений.
