+++ js/form-data.js
@@ -1,0 +1,1 @@
+//TODO: Здесь будут данные для формы.
