+++ js/mark.js
@@ -1,0 +1,1 @@
+//TODO: Здесь будет реализована работа с меткой на карте.
