+++ js/restrictions.js
@@ -1,0 +1,1 @@
+//TODO: Здесь будет реализована работа с ограничениями для полей ввода.
