+++ js/server-data.js
@@ -1,0 +1,1 @@
+//TODO: Модуль для обмена данными по сети.
